Adding in instructions for deploying to glass fish.  See the read me file.
Adding in the 'always' log level, as in $log.always("The logging service has started").

Open the logging log event library and changed it so that files are logged relative to rails root only.  Without this change file names become horrifyingly long when run under a J2EE server as the domain path is part of the file structure.

Fixed the home page link to still work when the context root in a J2EE server is something other than '/'.

Fixed the javascript loading so that it can work when the context root is something other than '/'.

Code reviewed by Greg Bowman.